+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 - Added new publisher values for Cboe Titanium Cboe Global Indices Feed
+- Made `detail::Buffer` shifts explicit to avoid redundant moves during record decoding
 - Added `Year` to `SplitDuration` enum for yearly historical batch job submissions
 
 ## 0.54.0 - 2026-04-21